NOJIRA: Fixed small issues identified by FindBugs tool
git-svn-id: https://source.sakaiproject.org/svn/sitestats/ufp/sitestats/trunk@72077 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ sitestats/sitestats-tool/src/java/org/sakaiproject/sitestats/tool/wicket/widget/ResourcesWidget.java
@@ -157,7 +157,7 @@ public class ResourcesWidget extends Panel {
 			
 			@Override
 			public String getSecondValue() {
-				double percentage = getTotalFiles()==0 ? 0 : round(100 * totalDistinctFileReads / getTotalFiles(), 0);
+				double percentage = getTotalFiles()==0 ? 0 : round(100 * totalDistinctFileReads / (double) getTotalFiles(), 0);
 				return String.valueOf((int) percentage) + '%';
 			}
